antes de deletar arquivo da pasta, porém funcionando o resto

+++ src/app.service.ts
@@ -12,6 +12,7 @@ import * as fs from 'fs/promises';
 import { User } from './interfaces/user.interface';
 import { UserAvatar } from './interfaces/UserAvatar';
 
+
 @Injectable()
 export class AppService {
   constructor(
@@ -50,69 +51,88 @@ export class AppService {
     return response.data;
   }
 
+
+
   async getUserAvatar(userId: string) {
     try {
+      let userAvatar: UserAvatar;
+  
       // Verificar se o usuário existe na API "https://reqres.in/"
       const userResponse = await axios.get(`https://reqres.in/api/users/${userId}`);
       const userData = userResponse.data.data;
-
-      if (!userData.avatar) {
-        throw new NotFoundException('User avatar not found.');
-      }
-
+  
       // Verificar se o avatar já foi salvo no banco de dados
-      let userAvatar: UserAvatar = await this.userAvatarModel.findOne({ userId }).exec();
-
+      userAvatar = await this.userAvatarModel.findOne({ userId }).exec();
+  
       if (!userAvatar) {
         // Se o avatar não estiver no banco de dados, obtenha-o da URL "avatar"
         const response = await axios.get(userData.avatar, {
           responseType: 'arraybuffer',
         });
-
+  
         // Converta o buffer da imagem em base64
         const base64Avatar = encode(new Uint8Array(response.data));
-
+  
         // Salve o avatar como um arquivo no diretório "avatars"
         const avatarFileName = `${userId}-${uuidv4()}.png`;
         const avatarPath = path.join(__dirname, '..', 'avatars', avatarFileName);
         await fs.writeFile(avatarPath, response.data);
-
-        // Salve o avatar no banco de dados
-        userAvatar = await this.userAvatarModel.create({
+  
+        // Salve o avatar no banco de dados com o campo "hash" definido corretamente
+        const userAvatarData = {
           userId,
-          hash: avatarFileName,
+          hash: avatarFileName, // Certifique-se de que a propriedade hash esteja correta
           avatar: base64Avatar,
-        });
+        };
+        console.log('User Avatar Data:', userAvatarData); // Verifique os dados do avatar antes de salvar
+        userAvatar = await this.userAvatarModel.create(userAvatarData);
       }
-
+  
       // Recupere a representação codificada em base64 do avatar do banco de dados
       return { avatar: userAvatar.avatar };
     } catch (error) {
       throw new NotFoundException('User not found.');
     }
   }
-
+  
   async deleteUserAvatar(userId: string) {
     try {
-      // Verificar se o usuário existe no banco de dados
-      const userAvatar: UserAvatar = await this.userAvatarModel.findOne({ userId }).exec();
-
+      const numericUserId = parseInt(userId, 10);
+  
+      const userAvatar: UserAvatar = await this.userAvatarModel.findOneAndDelete({ userId: numericUserId }).exec();
+      console.log("userAvatar == ", userAvatar);
+  
       if (!userAvatar) {
-        throw new NotFoundException('User avatar not found.');
+        return { message: 'Avatar not found.' };
       }
-
-      // Excluir o arquivo do armazenamento do sistema de arquivos (FileSystem)
-      const avatarPath = path.join(__dirname, '..', 'avatars', userAvatar.hash);
-      await fs.unlink(avatarPath);
-
-      // Remover a entrada do banco de dados
-      await this.userAvatarModel.deleteOne({ userId });
-
+  
+      if (!userAvatar.avatar) {
+        console.error('Error deleting avatar file: Avatar hash not found.');
+        throw new Error('An unexpected error occurred while deleting user avatar.');
+      }
+  
+      const avatarPath = path.join(__dirname, '..', 'avatars', userAvatar.avatar);
+      console.log("avatarPath = ", avatarPath);
+  
+      // Verificar se o arquivo do avatar existe antes de tentar excluí-lo
+      const fileExists = await fs.access(avatarPath)
+        .then(() => true)
+        .catch(() => false);
+  
+      if (fileExists) {
+        await fs.unlink(avatarPath);
+      } else {
+        console.error('Error deleting avatar file: Avatar file not found.');
+        throw new Error('An unexpected error occurred while deleting user avatar.');
+      }
+  
       return { message: 'Avatar deleted successfully.' };
     } catch (error) {
-      throw new NotFoundException('User not found.');
+      console.error('Error deleting user avatar:', error);
+      throw new Error('An unexpected error occurred while deleting user avatar.');
     }
   }
+  
 
   // Simula o envio de e-mail
   private sendEmail(user: User) {
